back end code to check for substrate-spec correctness moved to external page
