+++ server/routes/editSubstrateSpec.ts
@@ -1,11 +1,7 @@
-import { ApiPromise, WsProvider } from '@polkadot/api';
-import { RegisteredTypes } from '@polkadot/types/types';
 import { Request, Response, NextFunction } from 'express';
 import lookupCommunityIsVisibleToUser from '../util/lookupCommunityIsVisibleToUser';
 
-import { constructSubstrateUrl } from '../../shared/substrate';
-import { factory, formatFilename } from '../../shared/logging';
-const log = factory.getLogger(formatFilename(__filename));
+import testSubstrateSpec from '../util/testSubstrateSpec';
 
 const editSubstrateSpec = async (models, req: Request, res: Response, next: NextFunction) => {
   const [chain,, error] = await lookupCommunityIsVisibleToUser(models, req.body, req.user);
@@ -28,38 +24,10 @@ const editSubstrateSpec = async (models, req: Request, res: Response, next: Next
   });
   if (!requesterIsAdmin && !req.user.isAdmin) return next(new Error('Must be admin to edit'));
 
-  // get spec from request
-  let unpackedSpec: RegisteredTypes;
-  try {
-    unpackedSpec = JSON.parse(req.body.spec);
-  } catch (e) {
-    return next(new Error('Could not parse spec data'));
-  }
-  const sanitizedSpec: RegisteredTypes = {
-    types: unpackedSpec['types'],
-    typesAlias: unpackedSpec['typesAlias'],
-    typesBundle: unpackedSpec['typesBundle'],
-    typesChain: unpackedSpec['typesChain'],
-    typesSpec: unpackedSpec['typesSpec'],
-  };
-
-  // test out spec
   const nodes = await chain.getChainNodes();
   if (!nodes.length) return next(new Error('no chain nodes found'));
-  const provider = new WsProvider(constructSubstrateUrl(nodes[0].url), false);
-  try {
-    await provider.connect();
-  } catch (err) {
-    return next(new Error('failed to connect to node url'));
-  }
-  try {
-    const api = await ApiPromise.create({ provider, ...sanitizedSpec });
-    const version = api.runtimeVersion;
-    const props = await api.rpc.system.properties();
-    log.info(`Fetched version: ${version.specName}:${version.specVersion} and properties ${JSON.stringify(props)}`);
-  } catch (err) {
-    return next(new Error(`failed to initialize polkadot api: ${err.message}`));
-  }
+
+  const sanitizedSpec = await testSubstrateSpec(req.body.spec, nodes[0].url, next);
 
   // write back to database
   chain.substrate_spec = sanitizedSpec;
@@ -67,5 +35,6 @@ const editSubstrateSpec = async (models, req: Request, res: Response, next: Next
 
   return res.json({ status: 'Success', result: chain.toJSON() });
 };
+
 
 export default editSubstrateSpec;
